refactor: move some simple structs to obligation/primitives
