+++ core/credit/src/obligation/entity.rs
@@ -7,68 +7,7 @@ use es_entity::*;
 
 use crate::{primitives::*, CreditFacilityId};
 
-use super::error::ObligationError;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-pub enum ObligationStatus {
-    NotYetDue,
-    Due,
-    Overdue,
-    Defaulted,
-    Paid,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub struct ObligationAccounts {
-    pub receivable_account_id: CalaAccountId,
-    pub account_to_be_credited_id: CalaAccountId,
-}
-
-pub struct ObligationDueReallocationData {
-    pub tx_id: LedgerTxId,
-    pub amount: UsdCents,
-    pub not_yet_due_account_id: CalaAccountId,
-    pub due_account_id: CalaAccountId,
-}
-
-pub struct ObligationOverdueReallocationData {
-    pub tx_id: LedgerTxId,
-    pub outstanding_amount: UsdCents,
-    pub due_account_id: CalaAccountId,
-    pub overdue_account_id: CalaAccountId,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub struct ObligationsAmounts {
-    pub disbursed: UsdCents,
-    pub interest: UsdCents,
-}
-
-impl std::ops::Add<ObligationsAmounts> for ObligationsAmounts {
-    type Output = Self;
-
-    fn add(self, other: ObligationsAmounts) -> Self {
-        Self {
-            disbursed: self.disbursed + other.disbursed,
-            interest: self.interest + other.interest,
-        }
-    }
-}
-
-impl ObligationsAmounts {
-    pub const ZERO: Self = Self {
-        disbursed: UsdCents::ZERO,
-        interest: UsdCents::ZERO,
-    };
-
-    pub fn total(&self) -> UsdCents {
-        self.interest + self.disbursed
-    }
-
-    pub fn is_zero(&self) -> bool {
-        self.disbursed.is_zero() && self.interest.is_zero()
-    }
-}
+use super::{error::ObligationError, primitives::*};
 
 #[allow(clippy::large_enum_variant)]
 #[derive(EsEvent, Debug, Clone, Serialize, Deserialize)]
